[PR Feedback] Move LogLevel Enum to Constants
Goal:
Limit file imports for Public consumption

To limit the need to expose another class to consumers, this enum definition can be placed in constants with other enums for general purpose use.

Changes:
- Delete MMELogLevel.h
+ Move contents to Constants.h
+ Update imports
