+++ MapboxMobileEvents/MMEConstants.h
@@ -172,6 +172,33 @@ typedef NS_ENUM(NSInteger, MMEErrorNumber) {
     MMEErrorConfigUpdateError   = 10008
 };
 
+/*! Log Levels for Runtime Filtering */
+typedef NS_ENUM(NSUInteger, MMELogLevel) {
+    /// Log level for no messages
+    MMELogNone = 0,
+
+    /// Fatal Error Messages
+    MMELogFatal,
+
+    /// Error Messages
+    MMELogError,
+
+    /// Warning Messages
+    MMELogWarn,
+
+    /// Informational Messages
+    MMELogInfo,
+
+    /// Event Lifecycle Messages
+    MMELogEvent,
+
+    /// Network Connection Messages
+    MMELogNetwork,
+
+    /// All Debug Messages
+    MMELogDebug
+};
+
 /// key for MMEErrorEventInit userInfo dictionary containing the attributes which failed to create the event
 extern NSString * const MMEErrorEventAttributesKey;
