+++ MapboxMobileEvents/MMEEventsManager.h
@@ -1,7 +1,7 @@
 #import <Foundation/Foundation.h>
 #import <CoreLocation/CoreLocation.h>
 #import "MMETypes.h"
-#import "MMELogLevel.h"
+#import "MMEConstants.h"
 
 NS_ASSUME_NONNULL_BEGIN
