+++ MapboxMobileEvents/MMELogger.h
@@ -1,7 +1,7 @@
 #import "MMETypes.h"
 #import <Foundation/Foundation.h>
 #import <UIKit/UIKit.h>
-#import "MMELogLevel.h"
+#import "MMEConstants.h"
 
 NS_ASSUME_NONNULL_BEGIN
